Submit hack
So basically, auto detection of form submission is impossible for components inside shadow dom. So any button components cannot automattically trigger for submission if placed within a form and their type is set to submit. Ive created a work around to perform a form submission, but a better one will be needed.